refactor: split ui.rs into ui/ module directory
Split the monolithic ui.rs (1224 lines) into focused submodules:
- ui/mod.rs: re-exports, shared MenuIdMap type, find_menu_item helper
- ui/menu_builder.rs: menu construction (rebuild_tray_menu, build_device_submenu, etc.)
- ui/event_handler.rs: event handling (handle_menu_event, priority operations, etc.)

Each submodule has its own test block. All 82 tests pass.
